refactor: extract floating edit palette into dashboard-palette.ts
Move the edit-mode palette (draggable tile-kind chips + foyer toggle + its
drag-positioning) into a dashboard-palette component.

To avoid a render-timing trap, the palette registers its own chips as
gridstack drag-in sources in firstUpdated rather than the parent querying
them during grid init — a child component's chips don't exist yet when the
parent's updated() runs. The parent grid still accepts them via acceptWidgets
and owns the 'dropped' handler. Foyer changes surface as a bubbling, composed
foyer-toggled event that group-dashboard (draft state) and group-home (live
mirror) both observe.

group-dashboard.ts drops to ~1170 lines (from 1812). No behavior change
intended; palette drag-drop / move / foyer toggle need a manual edit-mode
check since the suite doesn't cover them.
