+++ src/renderer/src/groups/elements/group-dashboard.ts
@@ -36,6 +36,7 @@ import { AppOpenViews } from '../../layout/types.js';
 import { markdownParseSafe } from '../../utils.js';
 import { editIcon, closeIcon } from '../../ui/icons.js';
 import { dashboardStyles } from './dashboard-styles.js';
+import './dashboard-palette.js';
 import {
   DEFAULT_LAYOUTS,
   NewTileDraft,
@@ -98,9 +99,6 @@ export class GroupDashboardEl extends LitElement {
   @state() private _dialogEditingId: string | undefined = undefined;
   /** Draft foyer-enabled flag while editing; persisted on save. */
   @state() private _draftFoyerEnabled = true;
-  /** Floating palette position (.edit-layout-relative px). Undefined = the
-   * default bottom-left anchor. Session-only, not persisted. */
-  @state() private _palettePos: { left: number; top: number } | undefined;
   /**
    * When a palette chip is dropped onto the grid, the drop x/y/w/h is stashed
    * here so the ensuing add flow (dialog or picker) places the new tile at the
@@ -263,22 +261,17 @@ export class GroupDashboardEl extends LitElement {
     this._draftTiles =
       current && Array.isArray(current.tiles) ? structuredClone(current.tiles) : [];
     this._draftFoyerEnabled = current ? current.foyerEnabled !== false : true;
-    this._palettePos = undefined;
     this._editing = true;
     this._emitEditingChanged();
   }
 
-  /** Toggle the foyer on/off in the draft and tell group-home to reflect it
-   * live. Persisted with the dashboard on save. */
-  private _toggleFoyer(enabled: boolean) {
-    this._draftFoyerEnabled = enabled;
-    this.dispatchEvent(
-      new CustomEvent('foyer-toggled', {
-        detail: { enabled },
-        bubbles: true,
-        composed: true,
-      }),
-    );
+  /**
+   * The palette toggled the foyer. Reflect it in the draft (persisted on save);
+   * the palette's `foyer-toggled` event keeps bubbling past us to group-home,
+   * which mirrors the foyer live.
+   */
+  private _onFoyerToggled(e: CustomEvent<{ enabled: boolean }>) {
+    this._draftFoyerEnabled = e.detail.enabled;
   }
 
   /** Public: exit edit mode without saving. */
@@ -358,46 +351,8 @@ export class GroupDashboardEl extends LitElement {
       acceptWidgets: '.dash-palette-item',
     };
     this._grid = GridStack.init(opts, this._gridEl);
-    // Register the palette chips as drag-in sources. why: pass the actual
-    // elements, NOT a selector — setupDragIn's selector path uses
-    // document.querySelectorAll, which can't see the chips (they live in this
-    // component's light DOM inside group-home's shadow root). Query them from
-    // our own render root instead.
-    const chipEls = Array.from(
-      this.querySelectorAll<HTMLElement>('.dash-palette-item'),
-    );
-    if (chipEls.length) {
-      // why a custom helper: gridstack's `helper: 'clone'` deep-clones the
-      // chip and appends it to <body>. Some of our chip styling lives inside
-      // a <style> nested in the dashboard subtree and (depending on cascade)
-      // wasn't always applying to the clone — drag preview ended up as bare
-      // text. Snapshot the chip's real rendered box (computed background,
-      // border, radius, padding, font, dimensions) and inline those on a
-      // fresh element so the drag preview always looks like the chip itself.
-      const buildHelper = (el: HTMLElement): HTMLElement => {
-        const rect = el.getBoundingClientRect();
-        const cs = getComputedStyle(el);
-        const helper = el.cloneNode(true) as HTMLElement;
-        helper.style.boxSizing = 'border-box';
-        helper.style.width = `${rect.width}px`;
-        helper.style.height = `${rect.height}px`;
-        helper.style.background = cs.background;
-        helper.style.border = cs.border;
-        helper.style.borderLeft = cs.borderLeft;
-        helper.style.borderRadius = cs.borderRadius;
-        helper.style.padding = cs.padding;
-        helper.style.color = cs.color;
-        helper.style.font = cs.font;
-        helper.style.display = 'flex';
-        helper.style.alignItems = 'center';
-        helper.style.justifyContent = 'center';
-        helper.style.textAlign = 'center';
-        helper.style.opacity = '0.92';
-        helper.style.boxShadow = '0 6px 18px rgba(0,0,0,0.25)';
-        return helper;
-      };
-      GridStack.setupDragIn(chipEls, { appendTo: 'body', helper: buildHelper });
-    }
+    // The palette chips register themselves as drag-in sources (see
+    // dashboard-palette); this grid accepts them via acceptWidgets above.
     // When a chip is dropped, gridstack inserts a raw node at the drop point.
     // Capture its position + kind, remove that raw node, and route through the
     // normal add flow (dialog / picker) so the tile lands where it was dropped.
@@ -1070,89 +1025,13 @@ export class GroupDashboardEl extends LitElement {
     `;
   }
 
-  private _renderPalette() {
-    // why: chips carry gs-w/gs-h so the dropped widget gets a sensible default
-    // size, and data-kind so the dropped handler knows what tile to create.
-    const chip = (kind: DashboardTile['kind']) => ({
-      kind,
-      label: tileKindLabel(kind),
-      w: DEFAULT_LAYOUTS[kind].w,
-      h: DEFAULT_LAYOUTS[kind].h,
-    });
-    const chips = [chip('markdown'), chip('image'), chip('wal-embed'), chip('iframe')];
-    const pos = this._palettePos;
-    const posStyle = pos
-      ? `left:${pos.left}px; top:${pos.top}px; bottom:auto;`
-      : `left:12px; bottom:12px; top:auto;`;
-    return html`
-      <div class="dash-palette" style=${posStyle}>
-        <div class="dash-grip palette-grip" @pointerdown=${(e: PointerEvent) => this._startPaletteDrag(e)}>
-          <span class="dash-grip-dots">⠿</span>
-          <span class="palette-grip-label">${msg('Palette')}</span>
-        </div>
-        <div class="palette-body">
-          <div class="palette-section-title">${msg('Add')}</div>
-          ${chips.map(
-      (c) => html`
-              <div
-                class="dash-palette-item"
-                data-kind=${c.kind}
-                gs-w=${c.w}
-                gs-h=${c.h}
-                title=${msg('Drag onto the board')}
-              >
-                ${c.label}
-              </div>
-            `,
-    )}
-          <div class="palette-section-title">${msg('Options')}</div>
-          <label class="palette-option">
-            <input
-              type="checkbox"
-              .checked=${this._draftFoyerEnabled}
-              @change=${(e: Event) => this._toggleFoyer((e.target as HTMLInputElement).checked)}
-            />
-            ${msg('Foyer (chat)')}
-          </label>
-        </div>
-      </div>
-    `;
-  }
-
-  /**
-   * Drag the floating palette around by its grip. Positions are kept in
-   * .edit-layout-relative px (clamped inside it) and held only for the edit
-   * session (not persisted).
-   */
-  private _startPaletteDrag(e: PointerEvent) {
-    e.preventDefault();
-    const palette = (e.currentTarget as HTMLElement).closest('.dash-palette') as HTMLElement | null;
-    const container = this._editing ? this._gridEl?.closest('.edit-layout') : null;
-    if (!palette || !(container instanceof HTMLElement)) return;
-    const cRect = container.getBoundingClientRect();
-    const pRect = palette.getBoundingClientRect();
-    const grabDx = e.clientX - pRect.left;
-    const grabDy = e.clientY - pRect.top;
-    const move = (ev: PointerEvent) => {
-      let left = ev.clientX - cRect.left - grabDx;
-      let top = ev.clientY - cRect.top - grabDy;
-      // keep it within the container
-      left = Math.max(0, Math.min(left, cRect.width - pRect.width));
-      top = Math.max(0, Math.min(top, cRect.height - pRect.height));
-      this._palettePos = { left, top };
-    };
-    const up = () => {
-      window.removeEventListener('pointermove', move, true);
-      window.removeEventListener('pointerup', up, true);
-    };
-    window.addEventListener('pointermove', move, true);
-    window.addEventListener('pointerup', up, true);
-  }
-
   private _renderEditing() {
     return html`
       <div class="edit-layout">
-      ${this._renderPalette()}
+      <dashboard-palette
+        .foyerEnabled=${this._draftFoyerEnabled}
+        @foyer-toggled=${this._onFoyerToggled}
+      ></dashboard-palette>
       <div class="grid-stack" id="grid-root">
         ${repeat(
       this._draftTiles,
